+++ CHANGELOG.md
@@ -204,6 +204,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 - `Missing x-vtex-provider header` log.
 
+### Added
+
+- Aggregate documentation of this system on the README.
+
 ## [6.45.3] - 2021-08-30
 
 ### Fixed