+++ CHANGELOG.md
@@ -6,7 +6,11 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-## [7.0.0] - 2025-07-25
+## [7.0.1] - 2025-08-14
+### Changed
+- Update `diagnostics-nodejs` version, simplifying client initialization, consolidating telemetry logic, and improving type safety
+
+## [7.0.0] - 2025-08-13
 ### Changed
 - Version created to keep track of Runtime 7-compatible implementation of this lib